requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
discord-py==2.6.4
etelemetry==0.3.1
filelock==3.20.0
frozenlist==1.8.0
httplib2==0.31.0
idna==3.11
looseversion==1.3.0
lxml==6.0.2
multidict==6.7.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
propcache==0.4.1
prov==2.1.1
puremagic==1.30
py-cord==2.6.1
pydot==4.0.1
pymupdf==1.26.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyxnat==1.6.4
rdflib==7.4.0
requests==2.32.5
scipy==1.16.3
simplejson==3.20.2
six==1.17.0
soupsieve==2.8
traits==7.0.2
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
